+++ CHANGES.md
@@ -5,7 +5,9 @@
  - [ ] Build GitLab CI/CD Pipeline
  - [ ] Build GitHub CI/CD Pipeline
 
-## newproject v0.0.x
+## Beta version
 
- - [x] `0.0.2`: **`2020-12-28`**: `add`: Starting TDD, Sphinx documentation and pylint
- - [x] `0.0.1`: **`2020-04-18`**: `add`: Initial release
+ - [x] `2021.1.2b3`**: `add`: Added nox sessions for tests, coverage, linter,
+                              types, styles, notebooks and documentation
+ - [x] `2020.12.28b2`**: `add`: Starting TDD, Sphinx documentation and pylint
+ - [x] `2020.1.18b1`**: `add`: Initial release